Add the new terraform elasticsearch provider
This is the provider we use to run operations with terraform for our ES instances.
One of the current use cases is create templates for the index patterns we want
to have. This till now was happening manually with simple bash scripts or curl
commands.

Now it will be part of our terraform recipes.

Ticket: https://mattermost.atlassian.net/browse/MM-35109